initial 1.20.5 port

+++ neoforge/src/main/java/com/blamejared/crafttweaker/natives/event/anvil/ExpandAnvilUpdateEvent.java
@@ -57,13 +57,13 @@ public class ExpandAnvilUpdateEvent {
     }
     
     @ZenCodeType.Getter("cost")
-    public static int getCost(AnvilUpdateEvent internal) {
+    public static long getCost(AnvilUpdateEvent internal) {
         
         return internal.getCost();
     }
     
     @ZenCodeType.Setter("cost")
-    public static void setCost(AnvilUpdateEvent internal, int cost) {
+    public static void setCost(AnvilUpdateEvent internal, long cost) {
         
         internal.setCost(cost);
     }
